requirements/dev.txt
# Everything the developer needs in addition to the production requirements
-r prod.txt

# Testing
pytest==2.9.0
WebTest==2.0.20

# Management script
Flask-Script==2.0.5

# Lint and code style
flake8==2.5.4
flake8-blind-except==0.1.0
flake8-debugger==1.4.0
flake8-docstrings==0.2.5
flake8-isort==1.2
flake8-quotes==0.2.4
isort==4.2.2
pep8-naming==0.3.3
